requirements.txt
psycopg2-binary>=2.9
prowler>=3.0
pyyaml>=6.0
boto3>=1.34
